fastmvc_media/pyproject.toml
[project]
name = "fastmvc_media"
version = "0.1.0"
description = "File upload (multipart/S3-style), image variants, presigned URLs for FastMVC."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.100.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "pytest-cov>=4.0.0", "pytest-asyncio>=0.21.0"]
storage = ["fastmvc_storage>=0.1.0"]
pillow = ["Pillow>=10.0.0"]

[project.urls]
Homepage = "https://github.com/your-org/fastmvc_media"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
